inital commit

+++ controllers/studentController.js
@@ -32,31 +32,16 @@ module.exports.viewProfile = async function (req, res) {
     }
     res.render('student/profile', {student, availableCourses})
 }
-
-//render add
-module.exports.renderAddForm = function (req, res) {
-    const student = {
-        first_name: '',
-        last_name: '',
-        grade_level: 9,
-    }
-    res.render('student/add', {student});
-}
-
-//add
-module.exports.addStudent = async function (req, res) {
-    const student = await Student.create({
-        first_name: req.body.first_name,
-        last_name: req.body.last_name,
-        grade_level: req.body.grade_level
-    });
-    res.redirect(`/students/profile/${student.id}`);
-}
-
 //render edit
 module.exports.renderEditForm = async function (req, res) {
     const isAdmin =req.user.can('edit student');
     const profileBelonsToUSer=req.user.can('edit self')&& req.user.matchesStudent(req.params.id);
+
+    //if they arent staff and they arent the student whose profile they are trying to view,redirect
+    if(!isAdmin&&!profileBelonsToUSer){
+        res.redirect('/')
+        return
+    }
     const student = await Student.findByPk(req.params.id);
     console.log(student);
     res.render('student/edit', {student});
@@ -64,6 +49,14 @@ module.exports.renderEditForm = async function (req, res) {
 
 //update
 module.exports.updateStudent = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToYser=req.user.can('edit self')&& req.user.matchesStudentid(req.params.id);
+
+    //if they arent staff and they arent the student whose profile they are trying to view,redirect
+    if(!isAdmin&&!profileBelongsToYser){
+        res.redirect('/')
+        return
+    }
     const student = await Student.update({
         first_name: req.body.first_name,
         last_name: req.body.last_name,
@@ -79,6 +72,10 @@ module.exports.updateStudent = async function (req, res) {
 
 //delete
 module.exports.deleteStudent = async function (req, res) {
+    if(!user.can('delete student')){
+        res.redirect('/')
+        return
+    }
     await Student.destroy({
         where: {
             id: req.params.id
@@ -89,7 +86,15 @@ module.exports.deleteStudent = async function (req, res) {
 
 //Add course to student
 module.exports.enrollStudent = async function (req, res) {
+    const isAdmin = req.user.can('enroll student');
+    const profileBelongsToUser=req.user.can('enroll self')&& req.user.matchesStudentid(req.params.studentid);
 
+
+    //if they arent staff anf they arent the student whose profile they are trying to view,redirect
+    if(!isAdmin&&!profileBelongsToUser){
+        res.redirect('/')
+        return
+    }
         await StudentCourses.create({
             student_id: req.params.studentId,
             course_id: req.body.course
@@ -100,6 +105,14 @@ module.exports.enrollStudent = async function (req, res) {
 
 //delete course from student
 module.exports.removeCourse = async function(req, res){
+    const isAdmin = req.user.can('drop student');
+    const profileBelongsToUser=req.user.can('drop self')&& req.user.matchesStudentid(req.params.studentid);
+
+    //if they arent staff anf they arent the student whose profile they are trying to view,redirect
+    if(!isAdmin&&!profileBelongsToUser){
+        res.redirect('/')
+        return
+    }
     await StudentCourses.destroy({
         where: {
             student_id: req.params.studentId,
